# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CompanyOS Core API",
    "description": "Phase 1.3 \u2014 dashboard BFF + workspace + auth + hello.",
    "license": {
      "name": "Proprietary",
      "identifier": "Proprietary"
    },
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/finance/webhooks/stripe": {
      "post": {
        "tags": [
          "finance-webhooks"
        ],
        "summary": "POST /api/v1/finance/webhooks/stripe",
        "operationId": "stripe_webhook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StripeWebhookFixture"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookAck"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "StripePaymentObject": {
        "type": "object",
        "required": [
          "id",
          "amount",
          "currency",
          "customer_id",
          "status"
        ],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "currency": {
            "type": "string"
          },
          "customer_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "invoice_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
      "StripeWebhookData": {
        "type": "object",
        "required": [
          "object"
        ],
        "properties": {
          "object": {
            "$ref": "#/components/schemas/StripePaymentObject"
          }
        }
      },
      "StripeWebhookFixture": {
        "type": "object",
        "required": [
          "id",
          "type",
          "created",
          "data"
        ],
        "properties": {
          "created": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "$ref": "#/components/schemas/StripeWebhookData"
          },
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "WebhookAck": {
        "type": "object",
        "required": [
          "received",
          "duplicate"
        ],
        "properties": {
          "duplicate": {
            "type": "boolean"
          },
          "payment_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "received": {
            "type": "boolean"
          }
        }
      }
    }
  },
...
}